+++ app/blockstore.go
@@ -9,19 +9,29 @@ import (
 // AccountBlocksMap maps block hashes to account blocks
 type AccountBlocksMap map[string]*tradeblocks.AccountBlock
 
+// AccountTokenHeadMap maps an 'account:token' string to their head block
+type AccountTokenHeadMap map[string]*tradeblocks.AccountBlock
+
 // AccountChangeListener is called whenever an account block is added or changed
 type AccountChangeListener func(hash string, b *tradeblocks.AccountBlock)
+
+// VoteBlocksMap maps block hashes to vote blocks
+type VoteBlocksMap map[string]*tradeblocks.VoteBlock
 
 // BlockStore stores all of the local blockchains
 type BlockStore struct {
 	AccountChangeListener AccountChangeListener
 	AccountBlocks         AccountBlocksMap
+	AccountHeads          AccountTokenHeadMap
+	VoteBlocks            VoteBlocksMap
 }
 
 // NewBlockStore allocates and returns a new BlockStore
 func NewBlockStore() *BlockStore {
 	return &BlockStore{
 		AccountBlocks: make(AccountBlocksMap),
+		AccountHeads:  make(AccountTokenHeadMap),
+		VoteBlocks:    make(VoteBlocksMap),
 	}
 }
 
@@ -35,6 +45,7 @@ func (s *BlockStore) AddBlock(b *tradeblocks.AccountBlock) (string, error) {
 	}
 	h := b.Hash()
 	s.AccountBlocks[h] = b
+	s.AccountHeads[s.accountHeadKey(b.Account, b.Token)] = b
 	if s.AccountChangeListener != nil {
 		s.AccountChangeListener(h, b)
 	}
@@ -46,14 +57,14 @@ func (s *BlockStore) checkConflict(b *tradeblocks.AccountBlock) error {
 	if b.Previous == "" {
 		for _, block := range s.AccountBlocks {
 			if block.Previous == "" && block.Account == b.Account {
-				return &blockConflictError{block}
+				return &BlockConflictError{block}
 			}
 		}
 		return nil
 	}
 	for _, block := range s.AccountBlocks {
 		if block.Previous == b.Previous {
-			return &blockConflictError{block}
+			return &BlockConflictError{block}
 		}
 	}
 	return nil
@@ -63,6 +74,15 @@ func (s *BlockStore) checkConflict(b *tradeblocks.AccountBlock) error {
 // error return added for future proofing
 func (s *BlockStore) GetBlock(hash string) (*tradeblocks.AccountBlock, error) {
 	return s.AccountBlocks[hash], nil
+}
+
+// GetHeadBlock returns the head block for the specified account-token blockchain
+func (s *BlockStore) GetHeadBlock(account, token string) (*tradeblocks.AccountBlock, error) {
+	return s.AccountHeads[s.accountHeadKey(account, token)], nil
+}
+
+func (s *BlockStore) accountHeadKey(account, token string) string {
+	return account + ":" + token
 }
 
 // SwapBlocksMap maps block hashes to Swap blocks
@@ -171,11 +191,12 @@ func (s *OrderBlockStore) GetOrderBlock(hash string) (*tradeblocks.OrderBlock, e
 	return s.OrderBlocks[hash], nil
 }
 
-type blockConflictError struct {
+// BlockConflictError represents a conflict (multiple parent claim)
+type BlockConflictError struct {
 	existing *tradeblocks.AccountBlock
 }
 
-func (e *blockConflictError) Error() string {
+func (e *BlockConflictError) Error() string {
 	return fmt.Sprintf("blockstore: conflict with existing block '%s'", e.existing.Hash())
 }