+++ node/node.go
@@ -1,12 +1,15 @@
 package node
 
 import (
+	"fmt"
 	"github.com/jephir/tradeblocks"
 	"github.com/jephir/tradeblocks/app"
 	"github.com/jephir/tradeblocks/fs"
 	"github.com/jephir/tradeblocks/web"
 	"log"
 	"net/http"
+	"os"
+	"path/filepath"
 	"sync"
 )
 
@@ -16,6 +19,7 @@ type blockHashMap map[string]struct{}
 
 // Node represents a node in the TradeBlocks network
 type Node struct {
+	dir     string
 	store   *app.BlockStore
 	storage *fs.BlockStorage
 	client  *http.Client
@@ -29,10 +33,11 @@ type Node struct {
 // NewNode creates a new node that bootstraps from the specified URL. An error is returned if boostrapping fails.
 func NewNode(dir string) (n *Node, err error) {
 	store := app.NewBlockStore()
-	storage := fs.NewBlockStorage(store, dir)
+	storage := fs.NewBlockStorage(store, blocksDir(dir))
 	server := web.NewServer(store)
 	c := &http.Client{}
 	n = &Node{
+		dir:               dir,
 		store:             store,
 		storage:           storage,
 		client:            c,
@@ -40,12 +45,19 @@ func NewNode(dir string) (n *Node, err error) {
 		peers:             make(peerMap),
 		seenAccountBlocks: make(blockHashMap),
 	}
-	err = storage.Load()
+	err = n.initStorage()
 	if err != nil {
 		return
 	}
 	store.AccountChangeListener = n.accountChangeHandler()
 	return
+}
+
+func (n *Node) initStorage() error {
+	if err := os.MkdirAll(blocksDir(n.dir), 0700); err != nil {
+		return err
+	}
+	return n.storage.Load()
 }
 
 // Bootstrap registers with the specified server and downloads all blocks
@@ -113,29 +125,47 @@ func (n *Node) accountChangeHandler() app.AccountChangeListener {
 		if _, found := n.seenAccountBlocks[hash]; !found {
 			n.seenAccountBlocks[hash] = struct{}{}
 			for address := range n.peers {
-				c := web.NewClient(address)
-				r, err := c.NewPostAccountRequest(b)
-				if err != nil {
-					log.Print(err)
-					continue
-				}
-				//ss, _ := app.SerializeAccountBlock(b)
-				//log.Printf("node: sending %s to %s: %s", hash, address, ss)
-				res, err := n.client.Do(r)
-				if err != nil {
-					log.Print(err)
-					continue
-				}
-				if res.StatusCode != http.StatusOK {
-					log.Printf("node: unexpected status code %d", res.StatusCode)
-					continue
+				if err := n.broadcastAccountBlock(address, b); err != nil {
+					log.Println(err.Error())
 				}
 			}
 		}
 	}
 }
 
+func (n *Node) broadcastAccountBlock(address string, b *tradeblocks.AccountBlock) error {
+	c := web.NewClient(address)
+	r, err := c.NewPostAccountRequest(b)
+	if err != nil {
+		return err
+	}
+	//ss, _ := app.SerializeAccountBlock(b)
+	//log.Printf("node: sending %s to %s: %s", hash, address, ss)
+	res, err := n.client.Do(r)
+	if err != nil {
+		return err
+	}
+	if res.StatusCode != http.StatusOK {
+		return fmt.Errorf("node: unexpected status code %d", res.StatusCode)
+	}
+	return nil
+}
+
+func (n *Node) broadcastVote(address string, b *tradeblocks.AccountBlock) error {
+	// v := &tradeblocks.VoteBlock{
+	// 	Account: b.Account,
+	// 	Link: b.Hash(),
+	// 	Order: 0,
+	// 	Signature: "",
+	// }
+	return nil
+}
+
 // Sync flushes all unbroadcasted blocks to known peers
 func (n *Node) Sync() error {
 	return nil
+}
+
+func blocksDir(dir string) string {
+	return filepath.Join(dir, "blocks")
 }