+++ web/client.go
@@ -8,6 +8,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
+	"strings"
 
 	"github.com/jephir/tradeblocks/app"
 
@@ -65,6 +66,50 @@ func (c *Client) DecodeGetBlocksResponse(res *http.Response) (result app.Account
 	return
 }
 
+// NewGetAccountBlockRequest returns an http.Request to get an account block by hash
+func (c *Client) NewGetAccountBlockRequest(hash string) (r *http.Request, err error) {
+	hash = strings.TrimSpace(hash)
+	r, err = c.newRequest("GET", "/account", nil)
+	if err != nil {
+		return
+	}
+	q := r.URL.Query()
+	q.Add("hash", hash)
+	r.URL.RawQuery = q.Encode()
+	return
+}
+
+// DecodeGetAccountBlockResponse returns the result of a GetAccountHead request
+func (c *Client) DecodeGetAccountBlockResponse(res *http.Response, result *tradeblocks.AccountBlock) error {
+	if err := c.checkResponse(res); err != nil {
+		return err
+	}
+	return json.NewDecoder(res.Body).Decode(result)
+}
+
+// NewGetAccountHeadRequest returns an http.Request to get the head of an account-token chain
+func (c *Client) NewGetAccountHeadRequest(account, token string) (r *http.Request, err error) {
+	account = strings.TrimSpace(account)
+	token = strings.TrimSpace(token)
+	r, err = c.newRequest("GET", "/head", nil)
+	if err != nil {
+		return
+	}
+	q := r.URL.Query()
+	q.Add("account", account)
+	q.Add("token", token)
+	r.URL.RawQuery = q.Encode()
+	return
+}
+
+// DecodeGetAccountHeadResponse returns the result of a GetAccountHead request
+func (c *Client) DecodeGetAccountHeadResponse(res *http.Response, result *tradeblocks.AccountBlock) error {
+	if err := c.checkResponse(res); err != nil {
+		return err
+	}
+	return json.NewDecoder(res.Body).Decode(result)
+}
+
 func (c *Client) newRequest(method, path string, body io.Reader) (r *http.Request, err error) {
 	u, err := url.Parse(c.base)
 	if err != nil {
@@ -77,8 +122,11 @@ func (c *Client) newRequest(method, path string, body io.Reader) (r *http.Reques
 
 func (c *Client) checkResponse(res *http.Response) error {
 	if res.StatusCode != http.StatusOK {
-		errBody, _ := ioutil.ReadAll(res.Body)
-		return fmt.Errorf("unexpected status code %d\nError is:\n %s", res.StatusCode, errBody)
+		b, err := ioutil.ReadAll(res.Body)
+		if err != nil {
+			return err
+		}
+		return fmt.Errorf("client: unexpected status code %d: %s", res.StatusCode, string(b))
 	}
 	return nil
 }