+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - 2018-06-04
+
+### Added
+
+* Swap chain validation and tests
+* Order chain validation and tests
+* Block conflict detection
+* Vote (ordering) blocks
+* API endpoint to retrieve account-token heads
+
+### Changed
+
+* Block store account validation enabled
+* Unit tests now check for signed blocks
+
 ## 0.2.0 - 2018-05-27
 
 ### Added
@@ -15,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Push-based multi-node block synchronization
 * Signing and verification of account blocks
 * SSE endpoint for retrieving account information
-* Visual block explorer at http://localhost:8080
+* Visual block explorer at http://localhost:3000
 
 ### Changed